allow login prompting on silent redirection
Currently when the token is silently renewed the call to get a new token
will fail with the error `login_required` if the user's login session
has expired. In our environment the login request is handled with
credentials passed by the browser, without the need for user
intervention, and the redirection chain will successfully generate a
token if `settings.prompt ="none"` is omitted.

I expect this is a pretty unusual case so logins for silent requests are
only allowed if the new `silent_redirect_allow_login_prompt` option is
specified.